- transitioner: the format string for a DB query had %.15d instead of %.15e.     That produced a messed-up query that assigned garbage values to:         host_app_version.turnaround_var         host_app_version.turnaround_q         host_app_version.max_jobs_per_day         host_app_version.consecutive_valid     To repair these:         - set turnaround_var and turnaround_q to zero         - if max_jobs_per_day is outside of             (0..config.daily_result_quota)             set it to config.daily_result_quota         - if consecutive_valid is outside (0..1000), set it to zero     I added a script, html/ops/repair_21812.php, that does this;     if you ran server code between [21181] and [21812], run this script. - scheduler/transitioner: add <debug_quota> log flag - changed the build system to always use -Wall     (if we'd done this before, this bug wouldn't have happened) - fixed a bunch of other compile warnings
svn path=/trunk/boinc/; revision=21812

+++ sched/sched_score.cpp
@@ -171,7 +171,7 @@ void JOB_SET::add_job(JOB& job) {
         }
     }
 
-    if (jobs.size() == max_jobs) {
+    if ((int)jobs.size() == max_jobs) {
         JOB& worst_job = jobs.back();
         jobs.pop_back();
         ssp->wu_results[worst_job.index].state = WR_STATE_PRESENT;
